IndexedDB/idbfactory_cmp2.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBFactory.cmp() - invalid value</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#widl-IDBFactory-cmp-short-any-first-any-second">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>
<div id=log></div>
<script>
    test( function() {
        assert_throws({name: 'TypeError'}, function() {
            indexedDB.cmp();
        });
    }, "IDBFactory.cmp() - no argument.");

    test( function() {
        assert_throws("DataError", function() {
            indexedDB.cmp(null, null);
        });
    }, "IDBFactory.cmp(null, null) - null parameters are incorrect.");

    test( function() {
        assert_throws("DataError", function() {
            indexedDB.cmp(1, null);
        });
        assert_throws("DataError", function() {
            indexedDB.cmp(null, 1);
        });
    }, "IDBFactory.cmp(1, null / null, 1) - null parameter is incorrect.");

    test( function() {
        assert_throws("DataError", function() {
            indexedDB.cmp(NaN, NaN);
        });
    }, "IDBFactory.cmp(NaN, NaN) - NaN parameters are incorrect.");

    test( function() {
        assert_throws("DataError", function() {
            indexedDB.cmp(1, NaN);
        });
        assert_throws("DataError", function() {
            indexedDB.cmp(NaN, 1);
        });
    }, "IDBFactory.cmp(1, NaN / NaN, 1) - NaN parameter is incorrect.");
</script>
